requirements.txt
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
filelock==3.20.3
iniconfig==2.3.0
numpy==2.4.1
oracledb==3.4.1
packaging==26.0
pandas==3.0.0
platformdirs==4.5.1
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
pyodbc==5.3.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==80.10.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
virtualenv==20.36.1
